+++ RELEASE_NOTES.md
@@ -7,8 +7,9 @@ _Date TBD_
 * Add surveillance tables to the database
 * Add pending surveillance tables to the database
 * Add add meaningful_use_users column to certified_product & certified_product_details
+* Add 'Suspended by ONC' and 'Terminated by ONC' to certification_status table
 
---- 
+---
 
 ## Version 6.0.0
 _15 November 2016_